refactor: split metrics
